+++ stac_fastapi/extensions/stac_fastapi/extensions/core/aggregation/__init__.py
@@ -1,0 +1,5 @@
+"""Aggregation extension module."""
+
+from .aggregation import AggregationExtension
+
+__all__ = ["AggregationExtension"]